test(mcp): keep OpenAPI parser regressions with generator tests
